+++ tests/utils_test.py
@@ -107,20 +107,26 @@ class TestUtils(unittest.TestCase):
 
     def test_ensure_file_path_valid(self):
         self.assertEqual(
-            ensure_file_path_valid("examples/a-b.c/d f/hardcode.yml"),
-            os.path.join(os.getcwd(), "examples/a_b_c/d_f/hardcode.yml"),
+            ensure_file_path_valid(
+                os.path.join("examples", "a-b.c", "d f", "hardcode.yml")
+            ),
+            os.path.join(os.getcwd(), "examples", "a_b_c", "d_f", "hardcode.yml"),
         )
         self.assertEqual(
-            ensure_file_path_valid("1/2B/3.yml"),
-            os.path.join(os.getcwd(), "T1/T2B/T3.yml"),
+            ensure_file_path_valid(os.path.join("1", "2B", "3.yml")),
+            os.path.join(os.getcwd(), "T1", "T2B", "T3.yml"),
         )
         self.assertEqual(
-            ensure_file_path_valid("examples/a-b.c/2B/hardcode.yml"),
-            os.path.join(os.getcwd(), "examples/a_b_c/T2B/hardcode.yml"),
+            ensure_file_path_valid(
+                os.path.join("examples", "a-b.c", "2B", "hardcode.yml")
+            ),
+            os.path.join(os.getcwd(), "examples", "a_b_c", "T2B", "hardcode.yml"),
         )
         self.assertEqual(
-            ensure_file_path_valid("examples/postman_echo/request_methods/"),
-            os.path.join(os.getcwd(), "examples/postman_echo/request_methods"),
+            ensure_file_path_valid(
+                os.path.join("examples", "postman_echo", "request_methods")
+            ),
+            os.path.join(os.getcwd(), "examples", "postman_echo", "request_methods"),
         )
         self.assertEqual(
             ensure_file_path_valid(os.path.join(os.getcwd(), "test.yml")),
@@ -130,7 +136,8 @@ class TestUtils(unittest.TestCase):
             ensure_file_path_valid(os.getcwd()), os.getcwd(),
         )
         self.assertEqual(
-            ensure_file_path_valid(os.getcwd() + ".csv"), os.getcwd() + ".csv",
+            ensure_file_path_valid(os.path.join(os.getcwd(), "demo", ".csv")),
+            os.path.join(os.getcwd(), "demo", ".csv"),
         )
 
     def test_safe_dump_json(self):