+++ docs/CHANGELOG.md
@@ -5,6 +5,7 @@
 **Fixed**
 
 - fix: avoid '.csv' been converted to '_csv'
+- fix: convert har to JSON format testcase
 
 ## 3.0.12 (2020-06-14)