+++ docs/CHANGELOG.md
@@ -7,6 +7,7 @@
 - fix: avoid '.csv' been converted to '_csv'
 - fix: convert har to JSON format testcase
 - fix: missing ${var} handling in overriding config variables
+- fix: SyntaxError caused by quote in case of headers."Set-Cookie"
 
 ## 3.0.12 (2020-06-14)